refactor to separate component
